refactor(api): move usecase to evaluation context
