+++ api/tests/evaluation/unit/domain/usecases/remember-user-has-seen-assessment-instructions_test.js
@@ -1,5 +1,5 @@
-import { rememberUserHasSeenAssessmentInstructions } from '../../../../lib/domain/usecases/remember-user-has-seen-assessment-instructions.js';
-import { expect, sinon } from '../../../test-helper.js';
+import { rememberUserHasSeenAssessmentInstructions } from '../../../../../src/evaluation/domain/usecases/remember-user-has-seen-assessment-instructions.js';
+import { expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | UseCase | remember-user-has-seen-assessment-instructions', function () {
   let userRepository;
